+++ CHANGELOG.md
@@ -6,6 +6,8 @@ permalink: changelog.html
 
 ## [Unreleased]
 
+## [0.1.0] - 2025-02-02
+
 ### Added
 
 - Update the documentation for the project [#3](https://github.com/anirvanmandal/vscode-shortcut/issues/3)